Rename rootReducer filename to index, combine reducers
