+++ src/reducers/CarsReducer.js
@@ -1,4 +1,4 @@
-import { BUY_ITEM, REMOVE_FEATURE } from "./../constants/ActionTypes";
+import { BUY_ITEM, REMOVE_FEATURE } from "../constants/ActionTypes";
 
 const initialState = {
   additionalPrice: 0,
@@ -14,7 +14,7 @@ const initialState = {
     { id: 2, name: "Racing detail package", price: 1500 },
     { id: 3, name: "Premium sound system", price: 500 },
     { id: 4, name: "Rear spoiler", price: 250 }
-  ],
+  ]
 };  
 
 export const carsReducer = (state = initialState, action) => {
