+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import { carImagesReducer } from './carImagesReducer';
+import { carsReducer } from './carsReducer';
+
+export const rootReducer = combineReducers({
+  carImagesReducer,
+  carsReducer
+});
