--- src/reducers/rootReducer.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'react-redux';
-import { carImagesReducer } from './CarImagesReducer';
-import { carsReducer } from './CarsReducer';
-
-export const rootReducer = () => {
-  return combineReducers(
-    carImagesReducer,
-    carsReducer
-  )
-}
